% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectordb.R
\name{nearest}
\alias{nearest}
\title{Nearest neighbours of a query string in a store}
\usage{
nearest(store, text, k = 15L)
}
\arguments{
\item{store}{An \code{nli_store}.}

\item{text}{Query string.}

\item{k}{Number of neighbours to return.}
}
\value{
A data frame: \code{id}, \code{label}, \code{fittable}, \code{sim}, ordered by similarity.
}
\description{
Nearest neighbours of a query string in a store
}
