% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{nlmixr2libingest-package}
\alias{nlmixr2libingest}
\alias{nlmixr2libingest-package}
\title{nlmixr2libingest: token-efficient ingestion tooling for nlmixr2lib}
\description{
Corpus-scale tooling that makes extracting published popPK/PD models into
\pkg{nlmixr2lib} cheaper while preserving a strict source-trace quality
invariant. See \code{inst/design/ARCHITECTURE.md}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nlmixr2/nlmixr2libingest}
  \item Report bugs at \url{https://github.com/nlmixr2/nlmixr2libingest/issues}
}

}
\author{
\strong{Maintainer}: William Denney \email{wdenney@humanpredictions.com}

Authors:
\itemize{
  \item William Denney \email{wdenney@humanpredictions.com}
}

}
\keyword{internal}
