% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{register_db_path}
\alias{register_db_path}
\title{Default on-disk location of the cached register index (a DuckDB file)}
\usage{
register_db_path()
}
\value{
Absolute path; the parent directory is created on demand.
}
\description{
Default on-disk location of the cached register index (a DuckDB file)
}
