% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectordb.R
\name{save_store}
\alias{save_store}
\alias{load_store}
\title{Save / load a similarity store}
\usage{
save_store(store, path)

load_store(path)
}
\arguments{
\item{store}{An \code{nli_store}.}

\item{path}{File path.}
}
\description{
Save / load a similarity store
}
